server-side caching implemented

+++ src/services/postgres/PlaylistsService.js
@@ -3,11 +3,13 @@ const { Pool } = require('pg');
 const AuthorizationError = require('../../exceptions/AuthorizationError');
 const InvariantError = require('../../exceptions/InvariantError');
 const NotFoundError = require('../../exceptions/NotFoundError');
+const mapDBToModel = require('../../utils');
 
 class PlaylistsService {
-  constructor(collaborationsService) {
+  constructor(collaborationsService, cacheService) {
     this._pool = new Pool();
     this._collaborationsService = collaborationsService;
+    this._cacheService = cacheService;
   }
 
   async addPlaylist({ name, owner }) {
@@ -23,28 +25,37 @@ class PlaylistsService {
       throw new InvariantError('Playlist failed to add');
     }
 
+    await this._cacheService.delete(`playlists:${owner}`);
     return result.rows[0].id;
   }
 
   async getPlaylists(owner) {
-    const query = {
-      text: `
-      SELECT playlists.id, playlists.name, users.username FROM playlists 
-      LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
-      LEFT JOIN users ON users.id = playlists.owner
-      WHERE playlists.owner = $1 OR collaborations.user_id = $1
-      `,
-      values: [owner],
-    };
+    try {
+      const result = await this._cacheService.get(`playlists:${owner}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: `
+        SELECT playlists.id, playlists.name, users.username FROM playlists 
+        LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
+        LEFT JOIN users ON users.id = playlists.owner
+        WHERE playlists.owner = $1 OR collaborations.user_id = $1
+        `,
+        values: [owner],
+      };
 
-    const result = await this._pool.query(query);
+      const result = await this._pool.query(query);
+      const mappedResult = result.rows.map(mapDBToModel);
 
-    return result.rows;
+      await this._cacheService.set(`playlists:${owner}`, JSON.stringify(mappedResult));
+
+      return result.rows;
+    }
   }
 
   async deletePlaylistById(id) {
     const query = {
-      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
@@ -52,6 +63,9 @@ class PlaylistsService {
     if (!result.rowCount) {
       throw new NotFoundError('Playlist failed to delete. Id not found');
     }
+
+    const { owner } = result.rows[0];
+    await this._cacheService.delete(`playlists:${owner}`);
   }
 
   async verifyPlaylistOwner(id, owner) {
